backend/requirements.txt
# Core Web Framework
fastapi==0.121.2
python-dotenv==1.0.1
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# PDF Processing
PyMuPDF==1.26.6

# Image Processing
Pillow==12.0.0

# Azure AI Document Intelligence
azure-ai-documentintelligence==1.0.0

# Azure Blob Storage  (Step 2)
azure-storage-blob==12.23.0

# Machine Learning / YOLO  (local fallback when Azure creds are absent)
torch==2.9.1
ultralytics==8.3.228

# Required by ultralytics and other packages
numpy==2.2.6
